Specifically use UInt64 instead of UInt to prevent crash on 32-bit platforms

+++ Sources/MVTTools/VectorTileDecoder.swift
@@ -109,7 +109,7 @@ extension VectorTile {
                 return string
             }
             else if value.hasIntValue {
-                return Int(value.intValue)
+                return Int64(value.intValue)
             }
             else if value.hasBoolValue {
                 return value.boolValue
@@ -121,10 +121,10 @@ extension VectorTile {
                 return Double(value.floatValue)
             }
             else if value.hasUintValue {
-                return UInt(value.uintValue)
+                return UInt64(value.uintValue)
             }
             else if value.hasSintValue {
-                return Int(value.sintValue)
+                return Int64(value.sintValue)
             }
             else {
                 return ""
